update rustup to v1.25.1, pin urls
Updates the urls and hashes to match the current latest (as of 5 hours
ago) rustup release.

Additionally, the urls have been adjusted to include the rustup version.
This should mean the next time there's a rustup release, this chocolatey
package should continue to work.

This should hopefully mean updates can happen with a little less
urgency.